requirements.txt
torch
transformers
datasets
evaluate
sentencepiece
accelerate
tqdm
opencv-python
Pillow
